move apply-filters to menus/lib
